+++ Elements/src/Serialization/glTF/GltfExtensions.cs
@@ -141,6 +141,7 @@ namespace Elements.Serialization.glTF
 
             var matId = 0;
             var texId = 0;
+            var normalTexId = 0;
             var imageId = 0;
             var samplerId = 0;
 
@@ -200,47 +201,50 @@ namespace Elements.Serialization.glTF
                     {
                         var tex = new Texture();
                         textures.Add(tex);
-
-                        var image = new glTFLoader.Schema.Image();
-
-                        using (var ms = new MemoryStream())
-                        {
-                            // Flip the texture image vertically
-                            // to align with OpenGL convention.
-                            // 0,1  1,1
-                            // 0,0  1,0
-                            using (var texImage = SixLabors.ImageSharp.Image.Load(material.Texture))
-                            {
-                                PngMetadata meta = texImage.Metadata.GetPngMetadata();
-                                textureHasTransparency = meta.ColorType == PngColorType.RgbWithAlpha || meta.ColorType == PngColorType.GrayscaleWithAlpha;
-                                texImage.Mutate(x => x.Flip(FlipMode.Vertical));
-                                texImage.Save(ms, new SixLabors.ImageSharp.Formats.Png.PngEncoder());
-                            }
-                            var imageData = ms.ToArray();
-                            image.BufferView = AddBufferView(bufferViews, 0, buffer.Count, imageData.Length, null, null);
-                            buffer.AddRange(imageData);
-                        }
-
-                        while (buffer.Count % 4 != 0)
-                        {
-                            buffer.Add(0);
-                        }
-
-                        image.MimeType = glTFLoader.Schema.Image.MimeTypeEnum.image_png;
+                        var image = CreateImage(material.Texture, bufferViews, buffer, out textureHasTransparency);
                         tex.Source = imageId;
                         images.Add(image);
 
-                        var sampler = new Sampler();
-                        sampler.MagFilter = Sampler.MagFilterEnum.LINEAR;
-                        sampler.MinFilter = Sampler.MinFilterEnum.LINEAR;
-                        sampler.WrapS = Sampler.WrapSEnum.REPEAT;
-                        sampler.WrapT = Sampler.WrapTEnum.REPEAT;
+                        var sampler = CreateSampler();
                         tex.Sampler = samplerId;
                         samplers.Add(sampler);
 
                         textureDict.Add(material.Texture, texId);
 
                         texId++;
+                        imageId++;
+                        samplerId++;
+                    }
+                }
+
+                if (material.NormalTexture != null && File.Exists(material.NormalTexture))
+                {
+                    var ti = new MaterialNormalTextureInfo();
+                    m.NormalTexture = ti;
+                    ti.Index = normalTexId;
+                    ti.Scale = 1.0f;
+                    // Use the same texture coordinate as the 
+                    // base texture.
+                    ti.TexCoord = 0;
+
+                    if (textureDict.ContainsKey(material.NormalTexture))
+                    {
+                        ti.Index = textureDict[material.NormalTexture];
+                    }
+                    else
+                    {
+                        var tex = new Texture();
+                        textures.Add(tex);
+                        var image = CreateImage(material.NormalTexture, bufferViews, buffer, out _);
+                        tex.Source = imageId;
+                        images.Add(image);
+                        textureDict.Add(material.NormalTexture, normalTexId);
+
+                        var sampler = CreateSampler();
+                        tex.Sampler = samplerId;
+                        samplers.Add(sampler);
+
+                        normalTexId++;
                         imageId++;
                         samplerId++;
                     }
@@ -277,6 +281,47 @@ namespace Elements.Serialization.glTF
             }
 
             return materialDict;
+        }
+
+        private static glTFLoader.Schema.Image CreateImage(string path, List<BufferView> bufferViews, List<byte> buffer, out bool textureHasTransparency)
+        {
+            var image = new glTFLoader.Schema.Image();
+
+            using (var ms = new MemoryStream())
+            {
+                // Flip the texture image vertically
+                // to align with OpenGL convention.
+                // 0,1  1,1
+                // 0,0  1,0
+                using (var texImage = SixLabors.ImageSharp.Image.Load(path))
+                {
+                    PngMetadata meta = texImage.Metadata.GetPngMetadata();
+                    textureHasTransparency = meta.ColorType == PngColorType.RgbWithAlpha || meta.ColorType == PngColorType.GrayscaleWithAlpha;
+                    texImage.Mutate(x => x.Flip(FlipMode.Vertical));
+                    texImage.Save(ms, new SixLabors.ImageSharp.Formats.Png.PngEncoder());
+                }
+                var imageData = ms.ToArray();
+                image.BufferView = AddBufferView(bufferViews, 0, buffer.Count, imageData.Length, null, null);
+                buffer.AddRange(imageData);
+            }
+
+            while (buffer.Count % 4 != 0)
+            {
+                buffer.Add(0);
+            }
+
+            image.MimeType = glTFLoader.Schema.Image.MimeTypeEnum.image_png;
+            return image;
+        }
+
+        private static glTFLoader.Schema.Sampler CreateSampler()
+        {
+            var sampler = new Sampler();
+            sampler.MagFilter = Sampler.MagFilterEnum.LINEAR;
+            sampler.MinFilter = Sampler.MinFilterEnum.LINEAR;
+            sampler.WrapS = Sampler.WrapSEnum.REPEAT;
+            sampler.WrapT = Sampler.WrapTEnum.REPEAT;
+            return sampler;
         }
 
         internal static void AddLights(this Gltf gltf, List<Light> lights, List<Node> nodes)