+++ CHANGELOG.md
@@ -3,11 +3,19 @@
 ## 0.8.4
 
 ### Added
-- `Identity` static class
+- `Elements.Light`
+- `Elements.PointLight`
+- `Elements.SpotLight`
+- `Identity.AddAddOverrideIdentity`
+- `Polygon.PointInternal()`
+
+### Changed
+- `Elements.DirectionalLight` now inherits from `Elements.Light`.
 
 ## 0.8.3
 
 ### Added
+
 - `Profile.ToModelCurves()`
 - `Profile.Difference()`
 - `Profile.Intersection()`
@@ -30,7 +38,9 @@
 - `Random.NextColor()` and `Random.NextMaterial()`
 - `Validator.DisableValidationOnConstruction`
 - `Vector3.ComputeDefaultBasisVectors()`
+
 ### Changed
+
 - Make MeshConverter deserialization more flexible to accommodate a schema used in function `input_schema`.
 - Prevent the Polygon / Polyline constructors from throwing an exception on duplicate vertices, by removing duplicates automatically.
 - Make `Grid1d` and `Grid2d` serializable
@@ -38,11 +48,13 @@
 - All solids and csgs will now have planar texture coordinates.
 - Triangles are now validated to check for 3 distinct vertex positions.
 
-
 ### Fixed
+
 - Fixed a bug where `Polygon.UnionAll` was sometimes returning null when it shouldn't (Thanks @M-Juliani !)
 - Fixed [#517](https://github.com/hypar-io/Elements/issues/517)
 - Fixed a bug where Grid2d subcells would not split correctly with `SplitAtPoint`
+- Fixed [#528](https://github.com/hypar-io/Elements/issues/528)
+
 ## 0.8.2
 
 ### Changed