+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Changed
 - `Elements.DirectionalLight` now inherits from `Elements.Light`.
+- Adding glb elements to a model uses a cache rather than fetching the stream every time.
 
 ## 0.8.3