+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Elements.PointLight`
 - `Elements.SpotLight`
 - `Identity.AddOverrideIdentity`
+- `Material.NormalTexture`
 - `Polygon.PointInternal()`
 
 ### Changed